students/SmerdinaAnastasiya/task_02/src/requirements.txt
Flask>=2.1
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
Flask-Migrate>=4.0
python-dotenv>=1.0
gunicorn>=20.1.0
